correction erreurs syntaxe

+++ controllers/sauce.js
@@ -1,40 +1,67 @@
 //Import du modele Sauce pour utilisation
-const Sauce = require('../models/Sauce');
+const Sauce = require('../models/sauce');
+const fs = require('fs');
 
 //Export de la fonction de création de sauce
 exports.createSauce = (req, res) => {
     const sauceObject = JSON.parse(req.body.sauce);
-    delete req.body._id;
-    console.log (req.body);
-    const sauce = new Sauce ({
-        ...sauceObject
+    delete sauceObject._id;
+    delete sauceObject._userId;
+    const sauce = new Sauce({
+        ...sauceObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-    console.log(sauce)
+    
     sauce.save()
-    .then(() => res.status(201).json({ message: 'Sauce enregistrée !'}))
+    .then(() => res.status(201).json({ message: 'Sauce enregistrée !' }))
     .catch(error => res.status(400).json({ error }));
 };
 
 //Export de la fonction de récupération d'une seule sauce
 exports.getOneSauce = (req, res) => {
-    Sauce.findOne({ _id: req.params.id})
-    .then(() => res.status(200).json(sauce))
-    .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => res.status(200).json(sauce))
+    .catch(error => res.status(404).json({ error }));
 };
 
 //Export de la fonction de modification de sauce
 exports.modifySauce = (req, res) => {
-    Sauce.updateOne({ _id: req.params.id}, { ...req, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Sauce modifiée !'}))
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` 
+    } : { ...req.body };
+
+    delete sauceObject._userId;
+    Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+        if (sauce.userId != req.auth.userId) {
+            res.status(401).json({ message: 'Non autorisé !'});
+        } else {
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
+            .catch(error => res.status(401).json({ error }));
+        }
+    })
     .catch(error => res.status(400).json({ error }));
-};
-
-
+};    
+    
 //Export de la fonction de suppression de sauce
 exports.deleteSauce = (req, res) => {
-    Sauce.deleteOne({ _id: req.params.id})
-    .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
-    .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+    .then(sauce => {
+        if (sauce.userId != req.auth.userId) {
+            res.status(401).json({ message: 'Non autorisé !' });
+        } else {
+            const filename = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
+                .catch(error => res.status(401).json({ error }));
+            });           
+        }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 //Export de la fonction de récupération de toutes les sauces
